requirements.txt
fastapi==0.115.6
numpy==2.2.0
Pillow==11.0.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
uvicorn==0.34.0
pandas: 2.2.2
numpy: 1.26.4
matplotlib: 3.9.1
seaborn: 0.13.2
tensorflow: 2.17.0
fastapi: 0.115.4
uvicorn: 0.32.0
